Enhanced collision detection and new mobs
-Collision detection now checks to see if the overlapping pixels are
transparent. Currently only checks the mob for transparency; need to add
checking the player for transparency but the player isn't transparent
right now
-Added a fountain mob, which heals the player on collision. Updated the
mob class to allow the player to specify a string when creating a mob
object, and that string relates to the file name of the image used. This
should be changed to an enumerated list in a future update.
-bug fixes